backend/requirements.txt
fastapi
uvicorn
pydantic
motor      # async MongoDB driver
python-dotenv
